Feature#7 (#8)
* tmp commit

* code clean

* add sync functionality

* code clean

* add sync work

+++ Stingle/Classes/Common/Core/Network/Request/STSyncRequest.swift
@@ -10,7 +10,7 @@ import Foundation
 
 enum STSyncRequest {
     
-    case getUpdates(lastSeenTime: Int, lastTrashSeenTime: Int, lastAlbumsSeenTime: Int, lastAlbumFilesSeenTime: Int, lastDelSeenTime: Int, lastContactsSeenTime: Int)
+    case getUpdates(lastSeenTime: UInt64, lastTrashSeenTime: UInt64, lastAlbumsSeenTime: UInt64, lastAlbumFilesSeenTime: UInt64, lastDelSeenTime: UInt64, lastContactsSeenTime: UInt64)
 
 }
